@@ -18,63 +18,63 @@
     "format": "./node_modules/.bin/nx format:write",
     "format:write": "./node_modules/.bin/nx format:write",
     "format:check": "./node_modules/.bin/nx format:check",
-    "update": "ng update @nrwl/schematics",
+    "update": "ng update @nrwl/workspace",
     "update:check": "ng update",
     "workspace-schematic": "./node_modules/.bin/nx workspace-schematic",
     "dep-graph": "./node_modules/.bin/nx dep-graph",
     "help": "./node_modules/.bin/nx help",
-    "build:lib": "ng-packagr -p libs/ngx-owl-carousel-o/package.json"
+    "build:lib": "ng-packagr -p libs/ngx-owl-carousel-o/package.json",
+    "affected": "./node_modules/.bin/nx affected"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "6.0.1",
-    "@angular/common": "6.0.1",
-    "@angular/compiler": "6.0.1",
-    "@angular/core": "6.0.1",
-    "@angular/forms": "6.0.1",
-    "@angular/platform-browser": "6.0.1",
-    "@angular/platform-browser-dynamic": "6.0.1",
-    "@angular/router": "6.0.1",
-    "@ngrx/effects": "5.2.0",
-    "@ngrx/router-store": "5.2.0",
-    "@ngrx/store": "5.2.0",
-    "@ngrx/store-devtools": "5.2.0",
-    "@nrwl/nx": "6.1.0",
+    "@angular/animations": "8.1.3",
+    "@angular/common": "8.1.3",
+    "@angular/compiler": "8.1.3",
+    "@angular/core": "8.1.3",
+    "@angular/forms": "8.1.3",
+    "@angular/platform-browser": "8.1.3",
+    "@angular/platform-browser-dynamic": "8.1.3",
+    "@angular/router": "8.1.3",
+    "@ngrx/effects": "8.1.0",
+    "@ngrx/router-store": "8.1.0",
+    "@ngrx/store": "8.1.0",
+    "@ngrx/store-devtools": "8.1.0",
+    "@nrwl/angular": "8.3.0",
     "animate.css": "^3.7.0",
     "core-js": "^2.5.4",
-    "ng-packagr": "^4.0.0-rc.0",
-    "ngrx-store-freeze": "0.2.2",
-    "rxjs": "6.0.0",
+    "rxjs": "6.5.2",
     "rxjs-compat": "6.0.0",
-    "zone.js": "^0.8.26"
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.1",
-    "@angular-devkit/build-ng-packagr": "~0.6.1",
-    "@angular/cli": "6.0.1",
-    "@angular/compiler-cli": "6.0.1",
-    "@angular/language-service": "6.0.1",
-    "@ngrx/schematics": "5.2.0",
-    "@nrwl/schematics": "6.1.0",
-    "@types/jasmine": "~2.8.6",
+    "@angular-devkit/build-angular": "^0.801.1",
+    "@angular-devkit/build-ng-packagr": "~0.801.1",
+    "@angular/cli": "8.1.1",
+    "@angular/compiler-cli": "8.1.3",
+    "@angular/language-service": "8.1.3",
+    "@ngrx/schematics": "8.1.0",
+    "@nrwl/workspace": "8.3.0",
+    "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
-    "codelyzer": "~4.2.1",
-    "jasmine-core": "~2.99.1",
-    "jasmine-marbles": "^0.3.1",
+    "codelyzer": "^5.0.0",
+    "dotenv": "6.2.0",
+    "jasmine-core": "~3.4.0",
+    "jasmine-marbles": "^0.6.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~1.4.2",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "ng-packagr": "^4.0.0-rc.0",
-    "prettier": "1.10.2",
-    "protractor": "~5.3.0",
-    "ts-node": "~5.0.1",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "ng-packagr": "^5.3.0",
+    "prettier": "1.15.2",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
     "tsickle": ">=0.25.5",
-    "tslib": "^1.7.1",
-    "tslint": "~5.9.1",
-    "typescript": "2.7.2"
+    "tslib": "^1.9.0",
+    "tslint": "~5.15.0",
+    "typescript": "~3.4.5"
   }
 }
