@@ -1,356 +1,1 @@
-/**
- * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
- */
-/**
- * @record
- */
-export function BreakpointOptions() { }
-if (false) {
-    /**
-     * The number of items you want to see on the screen.
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.items;
-    /**
-     * Infinity loop. Duplicate last and first items to get loop illusion.
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.loop;
-    /**
-     * Center item. Works well with even an odd number of items.
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.center;
-    /**
-     * Go backwards when the boundary has reached.
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.rewind;
-    /**
-     * Mouse drag.
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.mouseDrag;
-    /**
-     * Touch drag
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.touchDrag;
-    /**
-     * Stage pull to edge
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.pullDrag;
-    /**
-     * Item pull to edge.
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.freeDrag;
-    /**
-     * margin-right(px) on item
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.margin;
-    /**
-     * Padding left and right on stage (can see neighbours)
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.stagePadding;
-    /**
-     * Merge items. Works with \@Input option 'dataMerge' of CarouselSlideDirective.
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.merge;
-    /**
-     * Fit merged items if screen is smaller than items value.
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.mergeFit;
-    /**
-     * Set non grid content.  Works with \@Input option 'width' of CarouselSlideDirective.
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.autoWidth;
-    /**
-     * Speed Calculate while dragging
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.smartSpeed;
-    /**
-     * Speed Calculate while dragging
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.fluidSpeed;
-    /**
-     * Drag end speed
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.dragEndSpeed;
-    /**
-     * Responsive refresh rate. Period of time between firing 'resize' event and refreshing carousel.
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.responsiveRefreshRate;
-    /**
-     * Show next/prev buttons.
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.nav;
-    /**
-     * Navigation speed
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.navSpeed;
-    /**
-     * Navigation slide by x. 'page' string can be set to slide by page.
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.slideBy;
-    /**
-     * Show dots navigation
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.dots;
-    /**
-     * Show dots each x item
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.dotsEach;
-    /**
-     * Pagination speed.
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.dotsSpeed;
-    /**
-     * Autoplay.
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.autoplay;
-    /**
-     * Autoplay interval timeout
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.autoplayTimeout;
-    /**
-     * Autoplay speed
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.autoplaySpeed;
-    /**
-     * Class for CSS3 animation out.
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.animateOut;
-    /**
-     * Class for CSS3 animation in.
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.animateIn;
-    /**
-     * Enables changing carousel according to the biggest height of ass active slides providing slides have different height
-     * @type {?|undefined}
-     */
-    BreakpointOptions.prototype.autoHeight;
-}
-/**
- * @record
- */
-export function ResponsiveSettings() { }
-/**
- * @record
- */
-export function OwlOptions() { }
-if (false) {
-    /**
-     * The number of items you want to see on the screen.
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.items;
-    /**
-     * Infinity loop. Duplicate last and first items to get loop illusion.
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.loop;
-    /**
-     * Center item. Works well with even an odd number of items.
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.center;
-    /**
-     * Go backwards when the boundary has reached.
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.rewind;
-    /**
-     * Mouse drag.
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.mouseDrag;
-    /**
-     * Touch drag
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.touchDrag;
-    /**
-     * Stage pull to edge
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.pullDrag;
-    /**
-     * Item pull to edge.
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.freeDrag;
-    /**
-     * margin-right(px) on item
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.margin;
-    /**
-     * Padding left and right on stage (can see neighbours)
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.stagePadding;
-    /**
-     * Merge items. Works with \@Input option 'dataMerge' of CarouselSlideDirective.
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.merge;
-    /**
-     * Fit merged items if screen is smaller than items value.
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.mergeFit;
-    /**
-     * Set non grid content.  Works with \@Input option 'width' of CarouselSlideDirective.
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.autoWidth;
-    /**
-     * Start position
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.startPosition;
-    /**
-     * Change direction from Right to left
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.rtl;
-    /**
-     * Speed Calculate while dragging
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.smartSpeed;
-    /**
-     * Speed Calculate while dragging
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.fluidSpeed;
-    /**
-     * Drag end speed
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.dragEndSpeed;
-    /**
-     * Object containing responsive options. Can be set to false to remove responsive capabilities.
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.responsive;
-    /**
-     * Responsive refresh rate. Period of time between firing 'resize' event and refreshing carousel.
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.responsiveRefreshRate;
-    /**
-     * Show next/prev buttons.
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.nav;
-    /**
-     * Text in next/prev buttons. HTML allowed.
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.navText;
-    /**
-     * Navigation speed
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.navSpeed;
-    /**
-     * Navigation slide by x. 'page' string can be set to slide by page.
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.slideBy;
-    /**
-     * Show dots navigation
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.dots;
-    /**
-     * Show dots each x item
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.dotsEach;
-    /**
-     * Used with \@Input option 'dotContent' of CarouselSlideDirective.
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.dotsData;
-    /**
-     * Pagination speed.
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.dotsSpeed;
-    /**
-     * Autoplay.
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.autoplay;
-    /**
-     * Autoplay interval timeout
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.autoplayTimeout;
-    /**
-     * Pause on mouse hover
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.autoplayHoverPause;
-    /**
-     * Autoplay speed
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.autoplaySpeed;
-    /**
-     * Lazy loading of slides with images
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.lazyLoad;
-    /**
-     * Indicates how many items (slides with images) to the right (and left, when loop is true) will be pre-loaded
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.lazyLoadEager;
-    /**
-     * Class for CSS3 animation out.
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.animateOut;
-    /**
-     * Class for CSS3 animation in.
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.animateIn;
-    /**
-     * Enables changing carousel according to the biggest height of ass active slides providing slides have different height
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.autoHeight;
-    /**
-     * Enables listening to url hash (fragment) changes
-     * @type {?|undefined}
-     */
-    OwlOptions.prototype.URLhashListener;
-}
-//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoib3dsLW9wdGlvbnMubW9kZWwuanMiLCJzb3VyY2VSb290Ijoibmc6Ly9uZ3gtb3dsLWNhcm91c2VsLW8vIiwic291cmNlcyI6WyJsaWIvbW9kZWxzL293bC1vcHRpb25zLm1vZGVsLnRzIl0sIm5hbWVzIjpbXSwibWFwcGluZ3MiOiI7Ozs7Ozs7QUFDQSx1Q0F1SUM7Ozs7OztJQW5JQyxrQ0FBZTs7Ozs7SUFJZixpQ0FBZTs7Ozs7SUFJZixtQ0FBaUI7Ozs7O0lBSWpCLG1DQUFpQjs7Ozs7SUFLakIsc0NBQW9COzs7OztJQUlwQixzQ0FBb0I7Ozs7O0lBSXBCLHFDQUFtQjs7Ozs7SUFJbkIscUNBQW1COzs7OztJQUtuQixtQ0FBZ0I7Ozs7O0lBSWhCLHlDQUFzQjs7Ozs7SUFLdEIsa0NBQWdCOzs7OztJQUloQixxQ0FBbUI7Ozs7O0lBSW5CLHNDQUFvQjs7Ozs7SUFLcEIsdUNBQW9COzs7OztJQUlwQix1Q0FBcUI7Ozs7O0lBSXJCLHlDQUFnQzs7Ozs7SUFLaEMsa0RBQStCOzs7OztJQU0vQixnQ0FBYzs7Ozs7SUFJZCxxQ0FBNEI7Ozs7O0lBSTVCLG9DQUEwQjs7Ozs7SUFJMUIsaUNBQWU7Ozs7O0lBSWYscUNBQTRCOzs7OztJQUk1QixzQ0FBNkI7Ozs7O0lBTzdCLHFDQUFtQjs7Ozs7SUFLbkIsNENBQXlCOzs7OztJQUt6QiwwQ0FBaUM7Ozs7O0lBTWpDLHVDQUE4Qjs7Ozs7SUFLOUIsc0NBQTZCOzs7OztJQU03Qix1Q0FBcUI7Ozs7O0FBSXZCLHdDQUVDOzs7O0FBQ0QsZ0NBaUxDOzs7Ozs7SUE3S0MsMkJBQWU7Ozs7O0lBSWYsMEJBQWU7Ozs7O0lBSWYsNEJBQWlCOzs7OztJQUlqQiw0QkFBaUI7Ozs7O0lBS2pCLCtCQUFvQjs7Ozs7SUFJcEIsK0JBQW9COzs7OztJQUlwQiw4QkFBbUI7Ozs7O0lBSW5CLDhCQUFtQjs7Ozs7SUFLbkIsNEJBQWdCOzs7OztJQUloQixrQ0FBc0I7Ozs7O0lBS3RCLDJCQUFnQjs7Ozs7SUFJaEIsOEJBQW1COzs7OztJQUluQiwrQkFBb0I7Ozs7O0lBS3BCLG1DQUFnQzs7Ozs7SUFJaEMseUJBQWM7Ozs7O0lBS2QsZ0NBQW9COzs7OztJQUlwQixnQ0FBcUI7Ozs7O0lBSXJCLGtDQUFnQzs7Ozs7SUFLaEMsZ0NBQWdDOzs7OztJQUloQywyQ0FBK0I7Ozs7O0lBTS9CLHlCQUFjOzs7OztJQUlkLDZCQUFtQjs7Ozs7SUFJbkIsOEJBQTRCOzs7OztJQUk1Qiw2QkFBMEI7Ozs7O0lBSTFCLDBCQUFlOzs7OztJQUlmLDhCQUE0Qjs7Ozs7SUFJNUIsOEJBQW1COzs7OztJQUluQiwrQkFBNkI7Ozs7O0lBTzdCLDhCQUFtQjs7Ozs7SUFLbkIscUNBQXlCOzs7OztJQUt6Qix3Q0FBNkI7Ozs7O0lBSzdCLG1DQUFpQzs7Ozs7SUFNakMsOEJBQW1COzs7OztJQUtuQixtQ0FBdUI7Ozs7O0lBTXZCLGdDQUE4Qjs7Ozs7SUFLOUIsK0JBQTZCOzs7OztJQU03QixnQ0FBcUI7Ozs7O0lBTXJCLHFDQUEwQiIsInNvdXJjZXNDb250ZW50IjpbIlxyXG5leHBvcnQgaW50ZXJmYWNlIEJyZWFrcG9pbnRPcHRpb25zIHtcclxuICAvKipcclxuICAgKiBUaGUgbnVtYmVyIG9mIGl0ZW1zIHlvdSB3YW50IHRvIHNlZSBvbiB0aGUgc2NyZWVuLlxyXG4gICAqL1xyXG4gIGl0ZW1zPzogbnVtYmVyO1xyXG4gIC8qKlxyXG4gICAqIEluZmluaXR5IGxvb3AuIER1cGxpY2F0ZSBsYXN0IGFuZCBmaXJzdCBpdGVtcyB0byBnZXQgbG9vcCBpbGx1c2lvbi5cclxuICAgKi9cclxuICBsb29wPzogYm9vbGVhbjtcclxuICAvKipcclxuICAgKiBDZW50ZXIgaXRlbS4gV29ya3Mgd2VsbCB3aXRoIGV2ZW4gYW4gb2RkIG51bWJlciBvZiBpdGVtcy5cclxuICAgKi9cclxuICBjZW50ZXI/OiBib29sZWFuO1xyXG4gIC8qKlxyXG4gICAqIEdvIGJhY2t3YXJkcyB3aGVuIHRoZSBib3VuZGFyeSBoYXMgcmVhY2hlZC5cclxuICAgKi9cclxuICByZXdpbmQ/OiBib29sZWFuO1xyXG5cclxuICAvKipcclxuICAgKiBNb3VzZSBkcmFnLlxyXG4gICAqL1xyXG4gIG1vdXNlRHJhZz86IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogVG91Y2ggZHJhZ1xyXG4gICAqL1xyXG4gIHRvdWNoRHJhZz86IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogU3RhZ2UgcHVsbCB0byBlZGdlXHJcbiAgICovXHJcbiAgcHVsbERyYWc/OiBib29sZWFuO1xyXG4gIC8qKlxyXG4gICAqIEl0ZW0gcHVsbCB0byBlZGdlLlxyXG4gICAqL1xyXG4gIGZyZWVEcmFnPzogYm9vbGVhbjtcclxuXHJcbiAgLyoqXHJcbiAgICogbWFyZ2luLXJpZ2h0KHB4KSBvbiBpdGVtXHJcbiAgICovXHJcbiAgbWFyZ2luPzogbnVtYmVyO1xyXG4gIC8qKlxyXG4gICAqIFBhZGRpbmcgbGVmdCBhbmQgcmlnaHQgb24gc3RhZ2UgKGNhbiBzZWUgbmVpZ2hib3VycylcclxuICAgKi9cclxuICBzdGFnZVBhZGRpbmc/OiBudW1iZXI7XHJcblxyXG4gIC8qKlxyXG4gICAqIE1lcmdlIGl0ZW1zLiBXb3JrcyB3aXRoIEBJbnB1dCBvcHRpb24gJ2RhdGFNZXJnZScgb2YgQ2Fyb3VzZWxTbGlkZURpcmVjdGl2ZS5cclxuICAgKi9cclxuICBtZXJnZT86IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogRml0IG1lcmdlZCBpdGVtcyBpZiBzY3JlZW4gaXMgc21hbGxlciB0aGFuIGl0ZW1zIHZhbHVlLlxyXG4gICAqL1xyXG4gIG1lcmdlRml0PzogYm9vbGVhbjtcclxuICAvKipcclxuICAgKiBTZXQgbm9uIGdyaWQgY29udGVudC4gIFdvcmtzIHdpdGggQElucHV0IG9wdGlvbiAnd2lkdGgnIG9mIENhcm91c2VsU2xpZGVEaXJlY3RpdmUuXHJcbiAgICovXHJcbiAgYXV0b1dpZHRoPzogYm9vbGVhbjtcclxuXHJcbiAgLyoqXHJcbiAgICogU3BlZWQgQ2FsY3VsYXRlIHdoaWxlIGRyYWdnaW5nXHJcbiAgICovXHJcbiAgc21hcnRTcGVlZD86IG51bWJlcjtcclxuICAvKipcclxuICAgKiBTcGVlZCBDYWxjdWxhdGUgd2hpbGUgZHJhZ2dpbmdcclxuICAgKi9cclxuICBmbHVpZFNwZWVkPzogYm9vbGVhbjtcclxuICAvKipcclxuICAgKiBEcmFnIGVuZCBzcGVlZFxyXG4gICAqL1xyXG4gIGRyYWdFbmRTcGVlZD86IG51bWJlciB8IGJvb2xlYW47XHJcblxyXG4gIC8qKlxyXG4gICAqIFJlc3BvbnNpdmUgcmVmcmVzaCByYXRlLiBQZXJpb2Qgb2YgdGltZSBiZXR3ZWVuIGZpcmluZyAncmVzaXplJyBldmVudCBhbmQgcmVmcmVzaGluZyBjYXJvdXNlbC5cclxuICAgKi9cclxuICByZXNwb25zaXZlUmVmcmVzaFJhdGU/OiBudW1iZXI7XHJcblxyXG4gIC8vIE5hdmlnYXRpb25cclxuICAvKipcclxuICAgKiBTaG93IG5leHQvcHJldiBidXR0b25zLlxyXG4gICAqL1xyXG4gIG5hdj86IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogTmF2aWdhdGlvbiBzcGVlZFxyXG4gICAqL1xyXG4gIG5hdlNwZWVkPzogbnVtYmVyIHwgYm9vbGVhbjtcclxuICAvKipcclxuICAgKiBOYXZpZ2F0aW9uIHNsaWRlIGJ5IHguICdwYWdlJyBzdHJpbmcgY2FuIGJlIHNldCB0byBzbGlkZSBieSBwYWdlLlxyXG4gICAqL1xyXG4gIHNsaWRlQnk/OiBudW1iZXIgfCBzdHJpbmc7XHJcbiAgLyoqXHJcbiAgICogU2hvdyBkb3RzIG5hdmlnYXRpb25cclxuICAgKi9cclxuICBkb3RzPzogYm9vbGVhbjtcclxuICAvKipcclxuICAgKiBTaG93IGRvdHMgZWFjaCB4IGl0ZW1cclxuICAgKi9cclxuICBkb3RzRWFjaD86IG51bWJlciB8IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogUGFnaW5hdGlvbiBzcGVlZC5cclxuICAgKi9cclxuICBkb3RzU3BlZWQ/OiBudW1iZXIgfCBib29sZWFuO1xyXG5cclxuXHJcbiAgLy8gQVVUT1BMQVkgKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKlxyXG4gIC8qKlxyXG4gICAqIEF1dG9wbGF5LlxyXG4gICAqL1xyXG4gIGF1dG9wbGF5PzogYm9vbGVhbjtcclxuXHJcbiAgLyoqXHJcbiAgICogQXV0b3BsYXkgaW50ZXJ2YWwgdGltZW91dFxyXG4gICAqL1xyXG4gIGF1dG9wbGF5VGltZW91dD86IG51bWJlcjtcclxuXHJcbiAgLyoqXHJcbiAgICogQXV0b3BsYXkgc3BlZWRcclxuICAgKi9cclxuICBhdXRvcGxheVNwZWVkPzogbnVtYmVyIHwgYm9vbGVhbjtcclxuXHJcbiAgLy8gQU5JTUFURSAqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqXHJcbiAgLyoqXHJcbiAgICogQ2xhc3MgZm9yIENTUzMgYW5pbWF0aW9uIG91dC5cclxuICAgKi9cclxuICBhbmltYXRlT3V0Pzogc3RyaW5nIHwgYm9vbGVhbjtcclxuXHJcbiAgLyoqXHJcbiAgICogQ2xhc3MgZm9yIENTUzMgYW5pbWF0aW9uIGluLlxyXG4gICAqL1xyXG4gIGFuaW1hdGVJbj86IHN0cmluZyB8IGJvb2xlYW47XHJcblxyXG4gIC8vIEFVVE9IRUlHSFQgKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKlxyXG4gIC8qKlxyXG4gICAqIEVuYWJsZXMgY2hhbmdpbmcgY2Fyb3VzZWwgYWNjb3JkaW5nIHRvIHRoZSBiaWdnZXN0IGhlaWdodCBvZiBhc3MgYWN0aXZlIHNsaWRlcyBwcm92aWRpbmcgc2xpZGVzIGhhdmUgZGlmZmVyZW50IGhlaWdodFxyXG4gICAqL1xyXG4gIGF1dG9IZWlnaHQ/OiBib29sZWFuO1xyXG5cclxufVxyXG5cclxuZXhwb3J0IGludGVyZmFjZSBSZXNwb25zaXZlU2V0dGluZ3Mge1xyXG4gIFtrZXk6IG51bWJlcl06IEJyZWFrcG9pbnRPcHRpb25zO1xyXG59XHJcbmV4cG9ydCBpbnRlcmZhY2UgT3dsT3B0aW9ucyB7XHJcbiAgLyoqXHJcbiAgICogVGhlIG51bWJlciBvZiBpdGVtcyB5b3Ugd2FudCB0byBzZWUgb24gdGhlIHNjcmVlbi5cclxuICAgKi9cclxuICBpdGVtcz86IG51bWJlcjtcclxuICAvKipcclxuICAgKiBJbmZpbml0eSBsb29wLiBEdXBsaWNhdGUgbGFzdCBhbmQgZmlyc3QgaXRlbXMgdG8gZ2V0IGxvb3AgaWxsdXNpb24uXHJcbiAgICovXHJcbiAgbG9vcD86IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogQ2VudGVyIGl0ZW0uIFdvcmtzIHdlbGwgd2l0aCBldmVuIGFuIG9kZCBudW1iZXIgb2YgaXRlbXMuXHJcbiAgICovXHJcbiAgY2VudGVyPzogYm9vbGVhbjtcclxuICAvKipcclxuICAgKiBHbyBiYWNrd2FyZHMgd2hlbiB0aGUgYm91bmRhcnkgaGFzIHJlYWNoZWQuXHJcbiAgICovXHJcbiAgcmV3aW5kPzogYm9vbGVhbjtcclxuXHJcbiAgLyoqXHJcbiAgICogTW91c2UgZHJhZy5cclxuICAgKi9cclxuICBtb3VzZURyYWc/OiBib29sZWFuO1xyXG4gIC8qKlxyXG4gICAqIFRvdWNoIGRyYWdcclxuICAgKi9cclxuICB0b3VjaERyYWc/OiBib29sZWFuO1xyXG4gIC8qKlxyXG4gICAqIFN0YWdlIHB1bGwgdG8gZWRnZVxyXG4gICAqL1xyXG4gIHB1bGxEcmFnPzogYm9vbGVhbjtcclxuICAvKipcclxuICAgKiBJdGVtIHB1bGwgdG8gZWRnZS5cclxuICAgKi9cclxuICBmcmVlRHJhZz86IGJvb2xlYW47XHJcblxyXG4gIC8qKlxyXG4gICAqIG1hcmdpbi1yaWdodChweCkgb24gaXRlbVxyXG4gICAqL1xyXG4gIG1hcmdpbj86IG51bWJlcjtcclxuICAvKipcclxuICAgKiBQYWRkaW5nIGxlZnQgYW5kIHJpZ2h0IG9uIHN0YWdlIChjYW4gc2VlIG5laWdoYm91cnMpXHJcbiAgICovXHJcbiAgc3RhZ2VQYWRkaW5nPzogbnVtYmVyO1xyXG5cclxuICAvKipcclxuICAgKiBNZXJnZSBpdGVtcy4gV29ya3Mgd2l0aCBASW5wdXQgb3B0aW9uICdkYXRhTWVyZ2UnIG9mIENhcm91c2VsU2xpZGVEaXJlY3RpdmUuXHJcbiAgICovXHJcbiAgbWVyZ2U/OiBib29sZWFuO1xyXG4gIC8qKlxyXG4gICAqIEZpdCBtZXJnZWQgaXRlbXMgaWYgc2NyZWVuIGlzIHNtYWxsZXIgdGhhbiBpdGVtcyB2YWx1ZS5cclxuICAgKi9cclxuICBtZXJnZUZpdD86IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogU2V0IG5vbiBncmlkIGNvbnRlbnQuICBXb3JrcyB3aXRoIEBJbnB1dCBvcHRpb24gJ3dpZHRoJyBvZiBDYXJvdXNlbFNsaWRlRGlyZWN0aXZlLlxyXG4gICAqL1xyXG4gIGF1dG9XaWR0aD86IGJvb2xlYW47XHJcblxyXG4gIC8qKlxyXG4gICAqIFN0YXJ0IHBvc2l0aW9uXHJcbiAgICovXHJcbiAgc3RhcnRQb3NpdGlvbj86IG51bWJlciB8IHN0cmluZztcclxuICAvKipcclxuICAgKiBDaGFuZ2UgZGlyZWN0aW9uIGZyb20gUmlnaHQgdG8gbGVmdFxyXG4gICAqL1xyXG4gIHJ0bD86IGJvb2xlYW47XHJcblxyXG4gIC8qKlxyXG4gICAqIFNwZWVkIENhbGN1bGF0ZSB3aGlsZSBkcmFnZ2luZ1xyXG4gICAqL1xyXG4gIHNtYXJ0U3BlZWQ/OiBudW1iZXI7XHJcbiAgLyoqXHJcbiAgICogU3BlZWQgQ2FsY3VsYXRlIHdoaWxlIGRyYWdnaW5nXHJcbiAgICovXHJcbiAgZmx1aWRTcGVlZD86IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogRHJhZyBlbmQgc3BlZWRcclxuICAgKi9cclxuICBkcmFnRW5kU3BlZWQ/OiBudW1iZXIgfCBib29sZWFuO1xyXG5cclxuICAvKipcclxuICAgKiBPYmplY3QgY29udGFpbmluZyByZXNwb25zaXZlIG9wdGlvbnMuIENhbiBiZSBzZXQgdG8gZmFsc2UgdG8gcmVtb3ZlIHJlc3BvbnNpdmUgY2FwYWJpbGl0aWVzLlxyXG4gICAqL1xyXG4gIHJlc3BvbnNpdmU/OiBSZXNwb25zaXZlU2V0dGluZ3M7XHJcbiAgLyoqXHJcbiAgICogUmVzcG9uc2l2ZSByZWZyZXNoIHJhdGUuIFBlcmlvZCBvZiB0aW1lIGJldHdlZW4gZmlyaW5nICdyZXNpemUnIGV2ZW50IGFuZCByZWZyZXNoaW5nIGNhcm91c2VsLlxyXG4gICAqL1xyXG4gIHJlc3BvbnNpdmVSZWZyZXNoUmF0ZT86IG51bWJlcjtcclxuXHJcbiAgLy8gTmF2aWdhdGlvblxyXG4gIC8qKlxyXG4gICAqIFNob3cgbmV4dC9wcmV2IGJ1dHRvbnMuXHJcbiAgICovXHJcbiAgbmF2PzogYm9vbGVhbjtcclxuICAvKipcclxuICAgKiBUZXh0IGluIG5leHQvcHJldiBidXR0b25zLiBIVE1MIGFsbG93ZWQuXHJcbiAgICovXHJcbiAgbmF2VGV4dD86IHN0cmluZ1tdO1xyXG4gIC8qKlxyXG4gICAqIE5hdmlnYXRpb24gc3BlZWRcclxuICAgKi9cclxuICBuYXZTcGVlZD86IG51bWJlciB8IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogTmF2aWdhdGlvbiBzbGlkZSBieSB4LiAncGFnZScgc3RyaW5nIGNhbiBiZSBzZXQgdG8gc2xpZGUgYnkgcGFnZS5cclxuICAgKi9cclxuICBzbGlkZUJ5PzogbnVtYmVyIHwgc3RyaW5nO1xyXG4gIC8qKlxyXG4gICAqIFNob3cgZG90cyBuYXZpZ2F0aW9uXHJcbiAgICovXHJcbiAgZG90cz86IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogU2hvdyBkb3RzIGVhY2ggeCBpdGVtXHJcbiAgICovXHJcbiAgZG90c0VhY2g/OiBudW1iZXIgfCBib29sZWFuO1xyXG4gIC8qKlxyXG4gICAqIFVzZWQgd2l0aCBASW5wdXQgb3B0aW9uICdkb3RDb250ZW50JyBvZiBDYXJvdXNlbFNsaWRlRGlyZWN0aXZlLlxyXG4gICAqL1xyXG4gIGRvdHNEYXRhPzogYm9vbGVhbjtcclxuICAvKipcclxuICAgKiBQYWdpbmF0aW9uIHNwZWVkLlxyXG4gICAqL1xyXG4gIGRvdHNTcGVlZD86IG51bWJlciB8IGJvb2xlYW47XHJcblxyXG5cclxuICAvLyBBVVRPUExBWSAqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqXHJcbiAgLyoqXHJcbiAgICogQXV0b3BsYXkuXHJcbiAgICovXHJcbiAgYXV0b3BsYXk/OiBib29sZWFuO1xyXG5cclxuICAvKipcclxuICAgKiBBdXRvcGxheSBpbnRlcnZhbCB0aW1lb3V0XHJcbiAgICovXHJcbiAgYXV0b3BsYXlUaW1lb3V0PzogbnVtYmVyO1xyXG5cclxuICAvKipcclxuICAgKiBQYXVzZSBvbiBtb3VzZSBob3ZlclxyXG4gICAqL1xyXG4gIGF1dG9wbGF5SG92ZXJQYXVzZT86IGJvb2xlYW47XHJcblxyXG4gIC8qKlxyXG4gICAqIEF1dG9wbGF5IHNwZWVkXHJcbiAgICovXHJcbiAgYXV0b3BsYXlTcGVlZD86IG51bWJlciB8IGJvb2xlYW47XHJcblxyXG4gIC8vIExBWlkgTE9BRElORyAqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqXHJcbiAgLyoqXHJcbiAgICogTGF6eSBsb2FkaW5nIG9mIHNsaWRlcyB3aXRoIGltYWdlc1xyXG4gICAqL1xyXG4gIGxhenlMb2FkPzogYm9vbGVhbjtcclxuXHJcbiAgLyoqXHJcbiAgICogSW5kaWNhdGVzIGhvdyBtYW55IGl0ZW1zIChzbGlkZXMgd2l0aCBpbWFnZXMpIHRvIHRoZSByaWdodCAoYW5kIGxlZnQsIHdoZW4gbG9vcCBpcyB0cnVlKSB3aWxsIGJlIHByZS1sb2FkZWRcclxuICAgKi9cclxuICBsYXp5TG9hZEVhZ2VyPzogbnVtYmVyO1xyXG5cclxuICAvLyBBTklNQVRFICoqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKipcclxuICAvKipcclxuICAgKiBDbGFzcyBmb3IgQ1NTMyBhbmltYXRpb24gb3V0LlxyXG4gICAqL1xyXG4gIGFuaW1hdGVPdXQ/OiBzdHJpbmcgfCBib29sZWFuO1xyXG5cclxuICAvKipcclxuICAgKiBDbGFzcyBmb3IgQ1NTMyBhbmltYXRpb24gaW4uXHJcbiAgICovXHJcbiAgYW5pbWF0ZUluPzogc3RyaW5nIHwgYm9vbGVhbjtcclxuXHJcbiAgLy8gQVVUT0hFSUdIVCAqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqXHJcbiAgLyoqXHJcbiAgICogRW5hYmxlcyBjaGFuZ2luZyBjYXJvdXNlbCBhY2NvcmRpbmcgdG8gdGhlIGJpZ2dlc3QgaGVpZ2h0IG9mIGFzcyBhY3RpdmUgc2xpZGVzIHByb3ZpZGluZyBzbGlkZXMgaGF2ZSBkaWZmZXJlbnQgaGVpZ2h0XHJcbiAgICovXHJcbiAgYXV0b0hlaWdodD86IGJvb2xlYW47XHJcblxyXG4gIC8vIEhBU0ggKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKipcclxuICAvKipcclxuICAgKiBFbmFibGVzIGxpc3RlbmluZyB0byB1cmwgaGFzaCAoZnJhZ21lbnQpIGNoYW5nZXNcclxuICAgKi9cclxuICBVUkxoYXNoTGlzdGVuZXI/OiBib29sZWFuO1xyXG59XHJcbiJdfQ==
+//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoib3dsLW9wdGlvbnMubW9kZWwuanMiLCJzb3VyY2VSb290Ijoibmc6Ly9uZ3gtb3dsLWNhcm91c2VsLW8vIiwic291cmNlcyI6WyJsaWIvbW9kZWxzL293bC1vcHRpb25zLm1vZGVsLnRzIl0sIm5hbWVzIjpbXSwibWFwcGluZ3MiOiIiLCJzb3VyY2VzQ29udGVudCI6WyJcclxuZXhwb3J0IGludGVyZmFjZSBCcmVha3BvaW50T3B0aW9ucyB7XHJcbiAgLyoqXHJcbiAgICogVGhlIG51bWJlciBvZiBpdGVtcyB5b3Ugd2FudCB0byBzZWUgb24gdGhlIHNjcmVlbi5cclxuICAgKi9cclxuICBpdGVtcz86IG51bWJlcjtcclxuICAvKipcclxuICAgKiBJbmZpbml0eSBsb29wLiBEdXBsaWNhdGUgbGFzdCBhbmQgZmlyc3QgaXRlbXMgdG8gZ2V0IGxvb3AgaWxsdXNpb24uXHJcbiAgICovXHJcbiAgbG9vcD86IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogQ2VudGVyIGl0ZW0uIFdvcmtzIHdlbGwgd2l0aCBldmVuIGFuIG9kZCBudW1iZXIgb2YgaXRlbXMuXHJcbiAgICovXHJcbiAgY2VudGVyPzogYm9vbGVhbjtcclxuICAvKipcclxuICAgKiBHbyBiYWNrd2FyZHMgd2hlbiB0aGUgYm91bmRhcnkgaGFzIHJlYWNoZWQuXHJcbiAgICovXHJcbiAgcmV3aW5kPzogYm9vbGVhbjtcclxuXHJcbiAgLyoqXHJcbiAgICogTW91c2UgZHJhZy5cclxuICAgKi9cclxuICBtb3VzZURyYWc/OiBib29sZWFuO1xyXG4gIC8qKlxyXG4gICAqIFRvdWNoIGRyYWdcclxuICAgKi9cclxuICB0b3VjaERyYWc/OiBib29sZWFuO1xyXG4gIC8qKlxyXG4gICAqIFN0YWdlIHB1bGwgdG8gZWRnZVxyXG4gICAqL1xyXG4gIHB1bGxEcmFnPzogYm9vbGVhbjtcclxuICAvKipcclxuICAgKiBJdGVtIHB1bGwgdG8gZWRnZS5cclxuICAgKi9cclxuICBmcmVlRHJhZz86IGJvb2xlYW47XHJcblxyXG4gIC8qKlxyXG4gICAqIG1hcmdpbi1yaWdodChweCkgb24gaXRlbVxyXG4gICAqL1xyXG4gIG1hcmdpbj86IG51bWJlcjtcclxuICAvKipcclxuICAgKiBQYWRkaW5nIGxlZnQgYW5kIHJpZ2h0IG9uIHN0YWdlIChjYW4gc2VlIG5laWdoYm91cnMpXHJcbiAgICovXHJcbiAgc3RhZ2VQYWRkaW5nPzogbnVtYmVyO1xyXG5cclxuICAvKipcclxuICAgKiBNZXJnZSBpdGVtcy4gV29ya3Mgd2l0aCBASW5wdXQgb3B0aW9uICdkYXRhTWVyZ2UnIG9mIENhcm91c2VsU2xpZGVEaXJlY3RpdmUuXHJcbiAgICovXHJcbiAgbWVyZ2U/OiBib29sZWFuO1xyXG4gIC8qKlxyXG4gICAqIEZpdCBtZXJnZWQgaXRlbXMgaWYgc2NyZWVuIGlzIHNtYWxsZXIgdGhhbiBpdGVtcyB2YWx1ZS5cclxuICAgKi9cclxuICBtZXJnZUZpdD86IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogU2V0IG5vbiBncmlkIGNvbnRlbnQuICBXb3JrcyB3aXRoIEBJbnB1dCBvcHRpb24gJ3dpZHRoJyBvZiBDYXJvdXNlbFNsaWRlRGlyZWN0aXZlLlxyXG4gICAqL1xyXG4gIGF1dG9XaWR0aD86IGJvb2xlYW47XHJcblxyXG4gIC8qKlxyXG4gICAqIFNwZWVkIENhbGN1bGF0ZSB3aGlsZSBkcmFnZ2luZ1xyXG4gICAqL1xyXG4gIHNtYXJ0U3BlZWQ/OiBudW1iZXI7XHJcbiAgLyoqXHJcbiAgICogU3BlZWQgQ2FsY3VsYXRlIHdoaWxlIGRyYWdnaW5nXHJcbiAgICovXHJcbiAgZmx1aWRTcGVlZD86IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogRHJhZyBlbmQgc3BlZWRcclxuICAgKi9cclxuICBkcmFnRW5kU3BlZWQ/OiBudW1iZXIgfCBib29sZWFuO1xyXG5cclxuICAvKipcclxuICAgKiBSZXNwb25zaXZlIHJlZnJlc2ggcmF0ZS4gUGVyaW9kIG9mIHRpbWUgYmV0d2VlbiBmaXJpbmcgJ3Jlc2l6ZScgZXZlbnQgYW5kIHJlZnJlc2hpbmcgY2Fyb3VzZWwuXHJcbiAgICovXHJcbiAgcmVzcG9uc2l2ZVJlZnJlc2hSYXRlPzogbnVtYmVyO1xyXG5cclxuICAvLyBOYXZpZ2F0aW9uXHJcbiAgLyoqXHJcbiAgICogU2hvdyBuZXh0L3ByZXYgYnV0dG9ucy5cclxuICAgKi9cclxuICBuYXY/OiBib29sZWFuO1xyXG4gIC8qKlxyXG4gICAqIE5hdmlnYXRpb24gc3BlZWRcclxuICAgKi9cclxuICBuYXZTcGVlZD86IG51bWJlciB8IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogTmF2aWdhdGlvbiBzbGlkZSBieSB4LiAncGFnZScgc3RyaW5nIGNhbiBiZSBzZXQgdG8gc2xpZGUgYnkgcGFnZS5cclxuICAgKi9cclxuICBzbGlkZUJ5PzogbnVtYmVyIHwgc3RyaW5nO1xyXG4gIC8qKlxyXG4gICAqIFNob3cgZG90cyBuYXZpZ2F0aW9uXHJcbiAgICovXHJcbiAgZG90cz86IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogU2hvdyBkb3RzIGVhY2ggeCBpdGVtXHJcbiAgICovXHJcbiAgZG90c0VhY2g/OiBudW1iZXIgfCBib29sZWFuO1xyXG4gIC8qKlxyXG4gICAqIFBhZ2luYXRpb24gc3BlZWQuXHJcbiAgICovXHJcbiAgZG90c1NwZWVkPzogbnVtYmVyIHwgYm9vbGVhbjtcclxuXHJcblxyXG4gIC8vIEFVVE9QTEFZICoqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKipcclxuICAvKipcclxuICAgKiBBdXRvcGxheS5cclxuICAgKi9cclxuICBhdXRvcGxheT86IGJvb2xlYW47XHJcblxyXG4gIC8qKlxyXG4gICAqIEF1dG9wbGF5IGludGVydmFsIHRpbWVvdXRcclxuICAgKi9cclxuICBhdXRvcGxheVRpbWVvdXQ/OiBudW1iZXI7XHJcblxyXG4gIC8qKlxyXG4gICAqIEF1dG9wbGF5IHNwZWVkXHJcbiAgICovXHJcbiAgYXV0b3BsYXlTcGVlZD86IG51bWJlciB8IGJvb2xlYW47XHJcblxyXG4gIC8vIEFOSU1BVEUgKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKlxyXG4gIC8qKlxyXG4gICAqIENsYXNzIGZvciBDU1MzIGFuaW1hdGlvbiBvdXQuXHJcbiAgICovXHJcbiAgYW5pbWF0ZU91dD86IHN0cmluZyB8IGJvb2xlYW47XHJcblxyXG4gIC8qKlxyXG4gICAqIENsYXNzIGZvciBDU1MzIGFuaW1hdGlvbiBpbi5cclxuICAgKi9cclxuICBhbmltYXRlSW4/OiBzdHJpbmcgfCBib29sZWFuO1xyXG5cclxuICAvLyBBVVRPSEVJR0hUICoqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKipcclxuICAvKipcclxuICAgKiBFbmFibGVzIGNoYW5naW5nIGNhcm91c2VsIGFjY29yZGluZyB0byB0aGUgYmlnZ2VzdCBoZWlnaHQgb2YgYXNzIGFjdGl2ZSBzbGlkZXMgcHJvdmlkaW5nIHNsaWRlcyBoYXZlIGRpZmZlcmVudCBoZWlnaHRcclxuICAgKi9cclxuICBhdXRvSGVpZ2h0PzogYm9vbGVhbjtcclxuXHJcbn1cclxuXHJcbmV4cG9ydCBpbnRlcmZhY2UgUmVzcG9uc2l2ZVNldHRpbmdzIHtcclxuICBba2V5OiBudW1iZXJdOiBCcmVha3BvaW50T3B0aW9ucztcclxufVxyXG5leHBvcnQgaW50ZXJmYWNlIE93bE9wdGlvbnMge1xyXG4gIC8qKlxyXG4gICAqIFRoZSBudW1iZXIgb2YgaXRlbXMgeW91IHdhbnQgdG8gc2VlIG9uIHRoZSBzY3JlZW4uXHJcbiAgICovXHJcbiAgaXRlbXM/OiBudW1iZXI7XHJcbiAgLyoqXHJcbiAgICogSW5maW5pdHkgbG9vcC4gRHVwbGljYXRlIGxhc3QgYW5kIGZpcnN0IGl0ZW1zIHRvIGdldCBsb29wIGlsbHVzaW9uLlxyXG4gICAqL1xyXG4gIGxvb3A/OiBib29sZWFuO1xyXG4gIC8qKlxyXG4gICAqIENlbnRlciBpdGVtLiBXb3JrcyB3ZWxsIHdpdGggZXZlbiBhbiBvZGQgbnVtYmVyIG9mIGl0ZW1zLlxyXG4gICAqL1xyXG4gIGNlbnRlcj86IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogR28gYmFja3dhcmRzIHdoZW4gdGhlIGJvdW5kYXJ5IGhhcyByZWFjaGVkLlxyXG4gICAqL1xyXG4gIHJld2luZD86IGJvb2xlYW47XHJcblxyXG4gIC8qKlxyXG4gICAqIE1vdXNlIGRyYWcuXHJcbiAgICovXHJcbiAgbW91c2VEcmFnPzogYm9vbGVhbjtcclxuICAvKipcclxuICAgKiBUb3VjaCBkcmFnXHJcbiAgICovXHJcbiAgdG91Y2hEcmFnPzogYm9vbGVhbjtcclxuICAvKipcclxuICAgKiBTdGFnZSBwdWxsIHRvIGVkZ2VcclxuICAgKi9cclxuICBwdWxsRHJhZz86IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogSXRlbSBwdWxsIHRvIGVkZ2UuXHJcbiAgICovXHJcbiAgZnJlZURyYWc/OiBib29sZWFuO1xyXG5cclxuICAvKipcclxuICAgKiBtYXJnaW4tcmlnaHQocHgpIG9uIGl0ZW1cclxuICAgKi9cclxuICBtYXJnaW4/OiBudW1iZXI7XHJcbiAgLyoqXHJcbiAgICogUGFkZGluZyBsZWZ0IGFuZCByaWdodCBvbiBzdGFnZSAoY2FuIHNlZSBuZWlnaGJvdXJzKVxyXG4gICAqL1xyXG4gIHN0YWdlUGFkZGluZz86IG51bWJlcjtcclxuXHJcbiAgLyoqXHJcbiAgICogTWVyZ2UgaXRlbXMuIFdvcmtzIHdpdGggQElucHV0IG9wdGlvbiAnZGF0YU1lcmdlJyBvZiBDYXJvdXNlbFNsaWRlRGlyZWN0aXZlLlxyXG4gICAqL1xyXG4gIG1lcmdlPzogYm9vbGVhbjtcclxuICAvKipcclxuICAgKiBGaXQgbWVyZ2VkIGl0ZW1zIGlmIHNjcmVlbiBpcyBzbWFsbGVyIHRoYW4gaXRlbXMgdmFsdWUuXHJcbiAgICovXHJcbiAgbWVyZ2VGaXQ/OiBib29sZWFuO1xyXG4gIC8qKlxyXG4gICAqIFNldCBub24gZ3JpZCBjb250ZW50LiAgV29ya3Mgd2l0aCBASW5wdXQgb3B0aW9uICd3aWR0aCcgb2YgQ2Fyb3VzZWxTbGlkZURpcmVjdGl2ZS5cclxuICAgKi9cclxuICBhdXRvV2lkdGg/OiBib29sZWFuO1xyXG5cclxuICAvKipcclxuICAgKiBTdGFydCBwb3NpdGlvblxyXG4gICAqL1xyXG4gIHN0YXJ0UG9zaXRpb24/OiBudW1iZXIgfCBzdHJpbmc7XHJcbiAgLyoqXHJcbiAgICogQ2hhbmdlIGRpcmVjdGlvbiBmcm9tIFJpZ2h0IHRvIGxlZnRcclxuICAgKi9cclxuICBydGw/OiBib29sZWFuO1xyXG5cclxuICAvKipcclxuICAgKiBTcGVlZCBDYWxjdWxhdGUgd2hpbGUgZHJhZ2dpbmdcclxuICAgKi9cclxuICBzbWFydFNwZWVkPzogbnVtYmVyO1xyXG4gIC8qKlxyXG4gICAqIFNwZWVkIENhbGN1bGF0ZSB3aGlsZSBkcmFnZ2luZ1xyXG4gICAqL1xyXG4gIGZsdWlkU3BlZWQ/OiBib29sZWFuO1xyXG4gIC8qKlxyXG4gICAqIERyYWcgZW5kIHNwZWVkXHJcbiAgICovXHJcbiAgZHJhZ0VuZFNwZWVkPzogbnVtYmVyIHwgYm9vbGVhbjtcclxuXHJcbiAgLyoqXHJcbiAgICogT2JqZWN0IGNvbnRhaW5pbmcgcmVzcG9uc2l2ZSBvcHRpb25zLiBDYW4gYmUgc2V0IHRvIGZhbHNlIHRvIHJlbW92ZSByZXNwb25zaXZlIGNhcGFiaWxpdGllcy5cclxuICAgKi9cclxuICByZXNwb25zaXZlPzogUmVzcG9uc2l2ZVNldHRpbmdzO1xyXG4gIC8qKlxyXG4gICAqIFJlc3BvbnNpdmUgcmVmcmVzaCByYXRlLiBQZXJpb2Qgb2YgdGltZSBiZXR3ZWVuIGZpcmluZyAncmVzaXplJyBldmVudCBhbmQgcmVmcmVzaGluZyBjYXJvdXNlbC5cclxuICAgKi9cclxuICByZXNwb25zaXZlUmVmcmVzaFJhdGU/OiBudW1iZXI7XHJcblxyXG4gIC8vIE5hdmlnYXRpb25cclxuICAvKipcclxuICAgKiBTaG93IG5leHQvcHJldiBidXR0b25zLlxyXG4gICAqL1xyXG4gIG5hdj86IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogVGV4dCBpbiBuZXh0L3ByZXYgYnV0dG9ucy4gSFRNTCBhbGxvd2VkLlxyXG4gICAqL1xyXG4gIG5hdlRleHQ/OiBzdHJpbmdbXTtcclxuICAvKipcclxuICAgKiBOYXZpZ2F0aW9uIHNwZWVkXHJcbiAgICovXHJcbiAgbmF2U3BlZWQ/OiBudW1iZXIgfCBib29sZWFuO1xyXG4gIC8qKlxyXG4gICAqIE5hdmlnYXRpb24gc2xpZGUgYnkgeC4gJ3BhZ2UnIHN0cmluZyBjYW4gYmUgc2V0IHRvIHNsaWRlIGJ5IHBhZ2UuXHJcbiAgICovXHJcbiAgc2xpZGVCeT86IG51bWJlciB8IHN0cmluZztcclxuICAvKipcclxuICAgKiBTaG93IGRvdHMgbmF2aWdhdGlvblxyXG4gICAqL1xyXG4gIGRvdHM/OiBib29sZWFuO1xyXG4gIC8qKlxyXG4gICAqIFNob3cgZG90cyBlYWNoIHggaXRlbVxyXG4gICAqL1xyXG4gIGRvdHNFYWNoPzogbnVtYmVyIHwgYm9vbGVhbjtcclxuICAvKipcclxuICAgKiBVc2VkIHdpdGggQElucHV0IG9wdGlvbiAnZG90Q29udGVudCcgb2YgQ2Fyb3VzZWxTbGlkZURpcmVjdGl2ZS5cclxuICAgKi9cclxuICBkb3RzRGF0YT86IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogUGFnaW5hdGlvbiBzcGVlZC5cclxuICAgKi9cclxuICBkb3RzU3BlZWQ/OiBudW1iZXIgfCBib29sZWFuO1xyXG5cclxuXHJcbiAgLy8gQVVUT1BMQVkgKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKlxyXG4gIC8qKlxyXG4gICAqIEF1dG9wbGF5LlxyXG4gICAqL1xyXG4gIGF1dG9wbGF5PzogYm9vbGVhbjtcclxuXHJcbiAgLyoqXHJcbiAgICogQXV0b3BsYXkgaW50ZXJ2YWwgdGltZW91dFxyXG4gICAqL1xyXG4gIGF1dG9wbGF5VGltZW91dD86IG51bWJlcjtcclxuXHJcbiAgLyoqXHJcbiAgICogUGF1c2Ugb24gbW91c2UgaG92ZXJcclxuICAgKi9cclxuICBhdXRvcGxheUhvdmVyUGF1c2U/OiBib29sZWFuO1xyXG5cclxuICAvKipcclxuICAgKiBBdXRvcGxheSBzcGVlZFxyXG4gICAqL1xyXG4gIGF1dG9wbGF5U3BlZWQ/OiBudW1iZXIgfCBib29sZWFuO1xyXG5cclxuICAvLyBMQVpZIExPQURJTkcgKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKlxyXG4gIC8qKlxyXG4gICAqIExhenkgbG9hZGluZyBvZiBzbGlkZXMgd2l0aCBpbWFnZXNcclxuICAgKi9cclxuICBsYXp5TG9hZD86IGJvb2xlYW47XHJcblxyXG4gIC8qKlxyXG4gICAqIEluZGljYXRlcyBob3cgbWFueSBpdGVtcyAoc2xpZGVzIHdpdGggaW1hZ2VzKSB0byB0aGUgcmlnaHQgKGFuZCBsZWZ0LCB3aGVuIGxvb3AgaXMgdHJ1ZSkgd2lsbCBiZSBwcmUtbG9hZGVkXHJcbiAgICovXHJcbiAgbGF6eUxvYWRFYWdlcj86IG51bWJlcjtcclxuXHJcbiAgLy8gQU5JTUFURSAqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqXHJcbiAgLyoqXHJcbiAgICogQ2xhc3MgZm9yIENTUzMgYW5pbWF0aW9uIG91dC5cclxuICAgKi9cclxuICBhbmltYXRlT3V0Pzogc3RyaW5nIHwgYm9vbGVhbjtcclxuXHJcbiAgLyoqXHJcbiAgICogQ2xhc3MgZm9yIENTUzMgYW5pbWF0aW9uIGluLlxyXG4gICAqL1xyXG4gIGFuaW1hdGVJbj86IHN0cmluZyB8IGJvb2xlYW47XHJcblxyXG4gIC8vIEFVVE9IRUlHSFQgKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKlxyXG4gIC8qKlxyXG4gICAqIEVuYWJsZXMgY2hhbmdpbmcgY2Fyb3VzZWwgYWNjb3JkaW5nIHRvIHRoZSBiaWdnZXN0IGhlaWdodCBvZiBhc3MgYWN0aXZlIHNsaWRlcyBwcm92aWRpbmcgc2xpZGVzIGhhdmUgZGlmZmVyZW50IGhlaWdodFxyXG4gICAqL1xyXG4gIGF1dG9IZWlnaHQ/OiBib29sZWFuO1xyXG5cclxuICAvLyBIQVNIICoqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqXHJcbiAgLyoqXHJcbiAgICogRW5hYmxlcyBsaXN0ZW5pbmcgdG8gdXJsIGhhc2ggKGZyYWdtZW50KSBjaGFuZ2VzXHJcbiAgICovXHJcbiAgVVJMaGFzaExpc3RlbmVyPzogYm9vbGVhbjtcclxufVxyXG4iXX0=
