@@ -1,37 +1,6 @@
 /**
- * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
- */
-/**
  * Data model for managing classes of .owl-stage DOM element
  */
 export class StageData {
 }
-if (false) {
-    /**
-     * Determines css-rule 'transform'
-     * @type {?}
-     */
-    StageData.prototype.transform;
-    /**
-     *  Determines css-rule 'transition'
-     * @type {?}
-     */
-    StageData.prototype.transition;
-    /**
-     *  Determines css-rule 'width'
-     * @type {?}
-     */
-    StageData.prototype.width;
-    /**
-     *  Determines css-rule 'padding-left'
-     * @type {?}
-     */
-    StageData.prototype.paddingL;
-    /**
-     *  Determines css-rule 'padding-right'
-     * @type {?}
-     */
-    StageData.prototype.paddingR;
-}
-//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoic3RhZ2UtZGF0YS5tb2RlbC5qcyIsInNvdXJjZVJvb3QiOiJuZzovL25neC1vd2wtY2Fyb3VzZWwtby8iLCJzb3VyY2VzIjpbImxpYi9tb2RlbHMvc3RhZ2UtZGF0YS5tb2RlbC50cyJdLCJuYW1lcyI6W10sIm1hcHBpbmdzIjoiOzs7Ozs7O0FBR0EsTUFBTSxPQUFPLFNBQVM7Q0FxQnJCOzs7Ozs7SUFqQkEsOEJBQWtCOzs7OztJQUlsQiwrQkFBbUI7Ozs7O0lBSW5CLDBCQUF1Qjs7Ozs7SUFJdkIsNkJBQTBCOzs7OztJQUkxQiw2QkFBMEIiLCJzb3VyY2VzQ29udGVudCI6WyIvKipcclxuICogRGF0YSBtb2RlbCBmb3IgbWFuYWdpbmcgY2xhc3NlcyBvZiAub3dsLXN0YWdlIERPTSBlbGVtZW50XHJcbiAqL1xyXG5leHBvcnQgY2xhc3MgU3RhZ2VEYXRhIHtcclxuXHQvKipcclxuXHQgKiBEZXRlcm1pbmVzIGNzcy1ydWxlICd0cmFuc2Zvcm0nXHJcblx0ICovXHJcblx0dHJhbnNmb3JtOiBzdHJpbmc7XHJcblx0LyoqXHJcblx0ICogIERldGVybWluZXMgY3NzLXJ1bGUgJ3RyYW5zaXRpb24nXHJcbiAgICovXHJcblx0dHJhbnNpdGlvbjogc3RyaW5nO1xyXG5cdC8qKlxyXG5cdCAqICBEZXRlcm1pbmVzIGNzcy1ydWxlICd3aWR0aCdcclxuICAgKi9cclxuXHR3aWR0aDogbnVtYmVyIHwgc3RyaW5nO1xyXG5cdC8qKlxyXG5cdCAqICBEZXRlcm1pbmVzIGNzcy1ydWxlICdwYWRkaW5nLWxlZnQnXHJcbiAgICovXHJcblx0cGFkZGluZ0w6IG51bWJlciB8IHN0cmluZztcclxuXHQvKipcclxuXHQgKiAgRGV0ZXJtaW5lcyBjc3MtcnVsZSAncGFkZGluZy1yaWdodCdcclxuICAgKi9cclxuXHRwYWRkaW5nUjogbnVtYmVyIHwgc3RyaW5nO1xyXG59Il19
+//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoic3RhZ2UtZGF0YS5tb2RlbC5qcyIsInNvdXJjZVJvb3QiOiJuZzovL25neC1vd2wtY2Fyb3VzZWwtby8iLCJzb3VyY2VzIjpbImxpYi9tb2RlbHMvc3RhZ2UtZGF0YS5tb2RlbC50cyJdLCJuYW1lcyI6W10sIm1hcHBpbmdzIjoiQUFBQTs7R0FFRztBQUNILE1BQU0sT0FBTyxTQUFTO0NBcUJyQiIsInNvdXJjZXNDb250ZW50IjpbIi8qKlxyXG4gKiBEYXRhIG1vZGVsIGZvciBtYW5hZ2luZyBjbGFzc2VzIG9mIC5vd2wtc3RhZ2UgRE9NIGVsZW1lbnRcclxuICovXHJcbmV4cG9ydCBjbGFzcyBTdGFnZURhdGEge1xyXG5cdC8qKlxyXG5cdCAqIERldGVybWluZXMgY3NzLXJ1bGUgJ3RyYW5zZm9ybSdcclxuXHQgKi9cclxuXHR0cmFuc2Zvcm06IHN0cmluZztcclxuXHQvKipcclxuXHQgKiAgRGV0ZXJtaW5lcyBjc3MtcnVsZSAndHJhbnNpdGlvbidcclxuICAgKi9cclxuXHR0cmFuc2l0aW9uOiBzdHJpbmc7XHJcblx0LyoqXHJcblx0ICogIERldGVybWluZXMgY3NzLXJ1bGUgJ3dpZHRoJ1xyXG4gICAqL1xyXG5cdHdpZHRoOiBudW1iZXIgfCBzdHJpbmc7XHJcblx0LyoqXHJcblx0ICogIERldGVybWluZXMgY3NzLXJ1bGUgJ3BhZGRpbmctbGVmdCdcclxuICAgKi9cclxuXHRwYWRkaW5nTDogbnVtYmVyIHwgc3RyaW5nO1xyXG5cdC8qKlxyXG5cdCAqICBEZXRlcm1pbmVzIGNzcy1ydWxlICdwYWRkaW5nLXJpZ2h0J1xyXG4gICAqL1xyXG5cdHBhZGRpbmdSOiBudW1iZXIgfCBzdHJpbmc7XHJcbn0iXX0=
