@@ -1,6 +1,2 @@
-/**
- * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
- */
 export { CarouselModule, CarouselComponent, CarouselSlideDirective, SlidesOutputData, OwlRouterLinkDirective, OwlRouterLinkWithHrefDirective } from './lib/carousel/carousel.module';
-//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoicHVibGljX2FwaS5qcyIsInNvdXJjZVJvb3QiOiJuZzovL25neC1vd2wtY2Fyb3VzZWwtby8iLCJzb3VyY2VzIjpbInB1YmxpY19hcGkudHMiXSwibmFtZXMiOltdLCJtYXBwaW5ncyI6Ijs7OztBQUFBLE9BQU8sRUFDTCxjQUFjLEVBQ2QsaUJBQWlCLEVBQ2pCLHNCQUFzQixFQUN0QixnQkFBZ0IsRUFDaEIsc0JBQXNCLEVBQ3RCLDhCQUE4QixFQUMvQixNQUFNLGdDQUFnQyxDQUFDIiwic291cmNlc0NvbnRlbnQiOlsiZXhwb3J0IHtcclxuICBDYXJvdXNlbE1vZHVsZSxcclxuICBDYXJvdXNlbENvbXBvbmVudCxcclxuICBDYXJvdXNlbFNsaWRlRGlyZWN0aXZlLFxyXG4gIFNsaWRlc091dHB1dERhdGEsXHJcbiAgT3dsUm91dGVyTGlua0RpcmVjdGl2ZSxcclxuICBPd2xSb3V0ZXJMaW5rV2l0aEhyZWZEaXJlY3RpdmVcclxufSBmcm9tICcuL2xpYi9jYXJvdXNlbC9jYXJvdXNlbC5tb2R1bGUnO1xyXG5cclxuZXhwb3J0IHtcclxuICBPd2xPcHRpb25zLCBSZXNwb25zaXZlU2V0dGluZ3MsIEJyZWFrcG9pbnRPcHRpb25zXHJcbn0gZnJvbSAnLi9saWIvbW9kZWxzL293bC1vcHRpb25zLm1vZGVsJzsiXX0=
+//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoicHVibGljX2FwaS5qcyIsInNvdXJjZVJvb3QiOiJuZzovL25neC1vd2wtY2Fyb3VzZWwtby8iLCJzb3VyY2VzIjpbInB1YmxpY19hcGkudHMiXSwibmFtZXMiOltdLCJtYXBwaW5ncyI6IkFBQUEsT0FBTyxFQUNMLGNBQWMsRUFDZCxpQkFBaUIsRUFDakIsc0JBQXNCLEVBQ3RCLGdCQUFnQixFQUNoQixzQkFBc0IsRUFDdEIsOEJBQThCLEVBQy9CLE1BQU0sZ0NBQWdDLENBQUMiLCJzb3VyY2VzQ29udGVudCI6WyJleHBvcnQge1xyXG4gIENhcm91c2VsTW9kdWxlLFxyXG4gIENhcm91c2VsQ29tcG9uZW50LFxyXG4gIENhcm91c2VsU2xpZGVEaXJlY3RpdmUsXHJcbiAgU2xpZGVzT3V0cHV0RGF0YSxcclxuICBPd2xSb3V0ZXJMaW5rRGlyZWN0aXZlLFxyXG4gIE93bFJvdXRlckxpbmtXaXRoSHJlZkRpcmVjdGl2ZVxyXG59IGZyb20gJy4vbGliL2Nhcm91c2VsL2Nhcm91c2VsLm1vZHVsZSc7XHJcblxyXG5leHBvcnQge1xyXG4gIE93bE9wdGlvbnMsIFJlc3BvbnNpdmVTZXR0aW5ncywgQnJlYWtwb2ludE9wdGlvbnNcclxufSBmcm9tICcuL2xpYi9tb2RlbHMvb3dsLW9wdGlvbnMubW9kZWwnOyJdfQ==
