@@ -1,128 +1,37 @@
 /**
- * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
- */
-/**
  * Data model for single navigation button
  */
-var /**
- * Data model for single navigation button
- */
-NavButton = /** @class */ (function () {
+var NavButton = /** @class */ (function () {
     function NavButton() {
     }
     return NavButton;
 }());
-/**
- * Data model for single navigation button
- */
 export { NavButton };
-if (false) {
-    /**
-     * Make button disabled by attaching class .disabled; it's being used when otpion loop=false
-     * @type {?}
-     */
-    NavButton.prototype.disabled;
-    /**
-     * Text for button; can be simple text or html-markup (e.g. <i class="fa fa-next">)
-     * @type {?}
-     */
-    NavButton.prototype.htmlText;
-}
 /**
  * Data model for  navigation block
  */
-var /**
- * Data model for  navigation block
- */
-NavData = /** @class */ (function () {
+var NavData = /** @class */ (function () {
     function NavData() {
     }
     return NavData;
 }());
-/**
- * Data model for  navigation block
- */
 export { NavData };
-if (false) {
-    /**
-     * Rurns on or turns off navigation block
-     * @type {?}
-     */
-    NavData.prototype.disabled;
-    /**
-     * Navigation button 'prev'
-     * @type {?}
-     */
-    NavData.prototype.prev;
-    /**
-     * Navigation button 'next'
-     * @type {?}
-     */
-    NavData.prototype.next;
-}
 /**
  * data model for single owl dot-button
  */
-var /**
- * data model for single owl dot-button
- */
-OwlSingeDot = /** @class */ (function () {
+var OwlSingeDot = /** @class */ (function () {
     function OwlSingeDot() {
     }
     return OwlSingeDot;
 }());
-/**
- * data model for single owl dot-button
- */
 export { OwlSingeDot };
-if (false) {
-    /**
-     * Id for dot button
-     * @type {?}
-     */
-    OwlSingeDot.prototype.id;
-    /**
-     * Makes dot active by attaching .active class to it
-     * @type {?}
-     */
-    OwlSingeDot.prototype.active;
-    /**
-     * Inner content of dot; can be html-markup
-     * @type {?}
-     */
-    OwlSingeDot.prototype.innerContent;
-    /**
-     * Enabled css-class which gives right presentaion of innerContent of dot.
-     * @type {?}
-     */
-    OwlSingeDot.prototype.showInnerContent;
-}
 /**
  * Data model for owl dot-block
  */
-var /**
- * Data model for owl dot-block
- */
-DotsData = /** @class */ (function () {
+var DotsData = /** @class */ (function () {
     function DotsData() {
     }
     return DotsData;
 }());
-/**
- * Data model for owl dot-block
- */
 export { DotsData };
-if (false) {
-    /**
-     * Turns on or turns off navigation block
-     * @type {?}
-     */
-    DotsData.prototype.disabled;
-    /**
-     * Array of dots
-     * @type {?}
-     */
-    DotsData.prototype.dots;
-}
-//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoibmF2aWdhdGlvbi1kYXRhLm1vZGVscy5qcyIsInNvdXJjZVJvb3QiOiJuZzovL25neC1vd2wtY2Fyb3VzZWwtby8iLCJzb3VyY2VzIjpbImxpYi9tb2RlbHMvbmF2aWdhdGlvbi1kYXRhLm1vZGVscy50cyJdLCJuYW1lcyI6W10sIm1hcHBpbmdzIjoiOzs7Ozs7O0FBR0E7Ozs7SUFBQTtJQVNBLENBQUM7SUFBRCxnQkFBQztBQUFELENBQUMsQUFURCxJQVNDOzs7Ozs7Ozs7O0lBTEEsNkJBQWtCOzs7OztJQUlsQiw2QkFBaUI7Ozs7O0FBS2xCOzs7O0lBQUE7SUFhQSxDQUFDO0lBQUQsY0FBQztBQUFELENBQUMsQUFiRCxJQWFDOzs7Ozs7Ozs7O0lBVEEsMkJBQWtCOzs7OztJQUlsQix1QkFBZ0I7Ozs7O0lBSWhCLHVCQUFnQjs7Ozs7QUFLakI7Ozs7SUFBQTtJQWlCQSxDQUFDO0lBQUQsa0JBQUM7QUFBRCxDQUFDLEFBakJELElBaUJDOzs7Ozs7Ozs7O0lBYkEseUJBQVc7Ozs7O0lBSVgsNkJBQWdCOzs7OztJQUloQixtQ0FBc0I7Ozs7O0lBSXRCLHVDQUEyQjs7Ozs7QUFLNUI7Ozs7SUFBQTtJQVNBLENBQUM7SUFBRCxlQUFDO0FBQUQsQ0FBQyxBQVRELElBU0M7Ozs7Ozs7Ozs7SUFMQSw0QkFBa0I7Ozs7O0lBSWxCLHdCQUFvQiIsInNvdXJjZXNDb250ZW50IjpbIi8qKlxyXG4gKiBEYXRhIG1vZGVsIGZvciBzaW5nbGUgbmF2aWdhdGlvbiBidXR0b25cclxuICovXHJcbmV4cG9ydCBjbGFzcyBOYXZCdXR0b24ge1xyXG5cdC8qKlxyXG5cdCAqIE1ha2UgYnV0dG9uIGRpc2FibGVkIGJ5IGF0dGFjaGluZyBjbGFzcyAuZGlzYWJsZWQ7IGl0J3MgYmVpbmcgdXNlZCB3aGVuIG90cGlvbiBsb29wPWZhbHNlXHJcblx0ICovXHJcblx0ZGlzYWJsZWQ6IGJvb2xlYW47XHJcblx0LyoqXHJcblx0ICogVGV4dCBmb3IgYnV0dG9uOyBjYW4gYmUgc2ltcGxlIHRleHQgb3IgaHRtbC1tYXJrdXAgKGUuZy4gPGkgY2xhc3M9XCJmYSBmYS1uZXh0XCI+KVxyXG5cdCAqL1xyXG5cdGh0bWxUZXh0OiBzdHJpbmc7XHJcbn1cclxuLyoqXHJcbiAqIERhdGEgbW9kZWwgZm9yICBuYXZpZ2F0aW9uIGJsb2NrXHJcbiAqL1xyXG5leHBvcnQgY2xhc3MgTmF2RGF0YSB7XHJcblx0LyoqXHJcblx0ICogUnVybnMgb24gb3IgdHVybnMgb2ZmIG5hdmlnYXRpb24gYmxvY2tcclxuXHQgKi9cclxuXHRkaXNhYmxlZDogYm9vbGVhbjtcclxuXHQvKipcclxuXHQgKiBOYXZpZ2F0aW9uIGJ1dHRvbiAncHJldidcclxuXHQgKi9cclxuXHRwcmV2OiBOYXZCdXR0b247XHJcblx0LyoqXHJcblx0ICogTmF2aWdhdGlvbiBidXR0b24gJ25leHQnXHJcblx0ICovXHJcblx0bmV4dDogTmF2QnV0dG9uO1xyXG59XHJcbi8qKlxyXG4gKiBkYXRhIG1vZGVsIGZvciBzaW5nbGUgb3dsIGRvdC1idXR0b25cclxuICovXHJcbmV4cG9ydCBjbGFzcyBPd2xTaW5nZURvdCB7XHJcblx0LyoqXHJcblx0ICogSWQgZm9yIGRvdCBidXR0b25cclxuXHQgKi9cclxuXHRpZDogc3RyaW5nO1xyXG5cdC8qKlxyXG5cdCAqIE1ha2VzIGRvdCBhY3RpdmUgYnkgYXR0YWNoaW5nIC5hY3RpdmUgY2xhc3MgdG8gaXRcclxuXHQgKi9cclxuXHRhY3RpdmU6IGJvb2xlYW47XHJcblx0LyoqXHJcblx0ICogSW5uZXIgY29udGVudCBvZiBkb3Q7IGNhbiBiZSBodG1sLW1hcmt1cFxyXG5cdCAqL1xyXG5cdGlubmVyQ29udGVudD86IHN0cmluZztcclxuXHQvKipcclxuXHQgKiBFbmFibGVkIGNzcy1jbGFzcyB3aGljaCBnaXZlcyByaWdodCBwcmVzZW50YWlvbiBvZiBpbm5lckNvbnRlbnQgb2YgZG90LlxyXG5cdCAqL1xyXG5cdHNob3dJbm5lckNvbnRlbnQ/OiBib29sZWFuO1xyXG59XHJcbi8qKlxyXG4gKiBEYXRhIG1vZGVsIGZvciBvd2wgZG90LWJsb2NrXHJcbiAqL1xyXG5leHBvcnQgY2xhc3MgRG90c0RhdGEge1xyXG5cdC8qKlxyXG5cdCAqIFR1cm5zIG9uIG9yIHR1cm5zIG9mZiBuYXZpZ2F0aW9uIGJsb2NrXHJcblx0ICovXHJcblx0ZGlzYWJsZWQ6IGJvb2xlYW47XHJcblx0LyoqXHJcblx0ICogQXJyYXkgb2YgZG90c1xyXG5cdCAqL1xyXG5cdGRvdHM6IE93bFNpbmdlRG90W107XHJcbn0iXX0=
+//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoibmF2aWdhdGlvbi1kYXRhLm1vZGVscy5qcyIsInNvdXJjZVJvb3QiOiJuZzovL25neC1vd2wtY2Fyb3VzZWwtby8iLCJzb3VyY2VzIjpbImxpYi9tb2RlbHMvbmF2aWdhdGlvbi1kYXRhLm1vZGVscy50cyJdLCJuYW1lcyI6W10sIm1hcHBpbmdzIjoiQUFBQTs7R0FFRztBQUNIO0lBQUE7SUFTQSxDQUFDO0lBQUQsZ0JBQUM7QUFBRCxDQUFDLEFBVEQsSUFTQzs7QUFDRDs7R0FFRztBQUNIO0lBQUE7SUFhQSxDQUFDO0lBQUQsY0FBQztBQUFELENBQUMsQUFiRCxJQWFDOztBQUNEOztHQUVHO0FBQ0g7SUFBQTtJQWlCQSxDQUFDO0lBQUQsa0JBQUM7QUFBRCxDQUFDLEFBakJELElBaUJDOztBQUNEOztHQUVHO0FBQ0g7SUFBQTtJQVNBLENBQUM7SUFBRCxlQUFDO0FBQUQsQ0FBQyxBQVRELElBU0MiLCJzb3VyY2VzQ29udGVudCI6WyIvKipcclxuICogRGF0YSBtb2RlbCBmb3Igc2luZ2xlIG5hdmlnYXRpb24gYnV0dG9uXHJcbiAqL1xyXG5leHBvcnQgY2xhc3MgTmF2QnV0dG9uIHtcclxuXHQvKipcclxuXHQgKiBNYWtlIGJ1dHRvbiBkaXNhYmxlZCBieSBhdHRhY2hpbmcgY2xhc3MgLmRpc2FibGVkOyBpdCdzIGJlaW5nIHVzZWQgd2hlbiBvdHBpb24gbG9vcD1mYWxzZVxyXG5cdCAqL1xyXG5cdGRpc2FibGVkOiBib29sZWFuO1xyXG5cdC8qKlxyXG5cdCAqIFRleHQgZm9yIGJ1dHRvbjsgY2FuIGJlIHNpbXBsZSB0ZXh0IG9yIGh0bWwtbWFya3VwIChlLmcuIDxpIGNsYXNzPVwiZmEgZmEtbmV4dFwiPilcclxuXHQgKi9cclxuXHRodG1sVGV4dDogc3RyaW5nO1xyXG59XHJcbi8qKlxyXG4gKiBEYXRhIG1vZGVsIGZvciAgbmF2aWdhdGlvbiBibG9ja1xyXG4gKi9cclxuZXhwb3J0IGNsYXNzIE5hdkRhdGEge1xyXG5cdC8qKlxyXG5cdCAqIFJ1cm5zIG9uIG9yIHR1cm5zIG9mZiBuYXZpZ2F0aW9uIGJsb2NrXHJcblx0ICovXHJcblx0ZGlzYWJsZWQ6IGJvb2xlYW47XHJcblx0LyoqXHJcblx0ICogTmF2aWdhdGlvbiBidXR0b24gJ3ByZXYnXHJcblx0ICovXHJcblx0cHJldjogTmF2QnV0dG9uO1xyXG5cdC8qKlxyXG5cdCAqIE5hdmlnYXRpb24gYnV0dG9uICduZXh0J1xyXG5cdCAqL1xyXG5cdG5leHQ6IE5hdkJ1dHRvbjtcclxufVxyXG4vKipcclxuICogZGF0YSBtb2RlbCBmb3Igc2luZ2xlIG93bCBkb3QtYnV0dG9uXHJcbiAqL1xyXG5leHBvcnQgY2xhc3MgT3dsU2luZ2VEb3Qge1xyXG5cdC8qKlxyXG5cdCAqIElkIGZvciBkb3QgYnV0dG9uXHJcblx0ICovXHJcblx0aWQ6IHN0cmluZztcclxuXHQvKipcclxuXHQgKiBNYWtlcyBkb3QgYWN0aXZlIGJ5IGF0dGFjaGluZyAuYWN0aXZlIGNsYXNzIHRvIGl0XHJcblx0ICovXHJcblx0YWN0aXZlOiBib29sZWFuO1xyXG5cdC8qKlxyXG5cdCAqIElubmVyIGNvbnRlbnQgb2YgZG90OyBjYW4gYmUgaHRtbC1tYXJrdXBcclxuXHQgKi9cclxuXHRpbm5lckNvbnRlbnQ/OiBzdHJpbmc7XHJcblx0LyoqXHJcblx0ICogRW5hYmxlZCBjc3MtY2xhc3Mgd2hpY2ggZ2l2ZXMgcmlnaHQgcHJlc2VudGFpb24gb2YgaW5uZXJDb250ZW50IG9mIGRvdC5cclxuXHQgKi9cclxuXHRzaG93SW5uZXJDb250ZW50PzogYm9vbGVhbjtcclxufVxyXG4vKipcclxuICogRGF0YSBtb2RlbCBmb3Igb3dsIGRvdC1ibG9ja1xyXG4gKi9cclxuZXhwb3J0IGNsYXNzIERvdHNEYXRhIHtcclxuXHQvKipcclxuXHQgKiBUdXJucyBvbiBvciB0dXJucyBvZmYgbmF2aWdhdGlvbiBibG9ja1xyXG5cdCAqL1xyXG5cdGRpc2FibGVkOiBib29sZWFuO1xyXG5cdC8qKlxyXG5cdCAqIEFycmF5IG9mIGRvdHNcclxuXHQgKi9cclxuXHRkb3RzOiBPd2xTaW5nZURvdFtdO1xyXG59Il19
