@@ -1,47 +1,10 @@
 /**
- * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
- */
-/**
  * Data model for managing classes of .owl-stage DOM element
  */
-var /**
- * Data model for managing classes of .owl-stage DOM element
- */
-StageData = /** @class */ (function () {
+var StageData = /** @class */ (function () {
     function StageData() {
     }
     return StageData;
 }());
-/**
- * Data model for managing classes of .owl-stage DOM element
- */
 export { StageData };
-if (false) {
-    /**
-     * Determines css-rule 'transform'
-     * @type {?}
-     */
-    StageData.prototype.transform;
-    /**
-     *  Determines css-rule 'transition'
-     * @type {?}
-     */
-    StageData.prototype.transition;
-    /**
-     *  Determines css-rule 'width'
-     * @type {?}
-     */
-    StageData.prototype.width;
-    /**
-     *  Determines css-rule 'padding-left'
-     * @type {?}
-     */
-    StageData.prototype.paddingL;
-    /**
-     *  Determines css-rule 'padding-right'
-     * @type {?}
-     */
-    StageData.prototype.paddingR;
-}
-//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoic3RhZ2UtZGF0YS5tb2RlbC5qcyIsInNvdXJjZVJvb3QiOiJuZzovL25neC1vd2wtY2Fyb3VzZWwtby8iLCJzb3VyY2VzIjpbImxpYi9tb2RlbHMvc3RhZ2UtZGF0YS5tb2RlbC50cyJdLCJuYW1lcyI6W10sIm1hcHBpbmdzIjoiOzs7Ozs7O0FBR0E7Ozs7SUFBQTtJQXFCQSxDQUFDO0lBQUQsZ0JBQUM7QUFBRCxDQUFDLEFBckJELElBcUJDOzs7Ozs7Ozs7O0lBakJBLDhCQUFrQjs7Ozs7SUFJbEIsK0JBQW1COzs7OztJQUluQiwwQkFBdUI7Ozs7O0lBSXZCLDZCQUEwQjs7Ozs7SUFJMUIsNkJBQTBCIiwic291cmNlc0NvbnRlbnQiOlsiLyoqXHJcbiAqIERhdGEgbW9kZWwgZm9yIG1hbmFnaW5nIGNsYXNzZXMgb2YgLm93bC1zdGFnZSBET00gZWxlbWVudFxyXG4gKi9cclxuZXhwb3J0IGNsYXNzIFN0YWdlRGF0YSB7XHJcblx0LyoqXHJcblx0ICogRGV0ZXJtaW5lcyBjc3MtcnVsZSAndHJhbnNmb3JtJ1xyXG5cdCAqL1xyXG5cdHRyYW5zZm9ybTogc3RyaW5nO1xyXG5cdC8qKlxyXG5cdCAqICBEZXRlcm1pbmVzIGNzcy1ydWxlICd0cmFuc2l0aW9uJ1xyXG4gICAqL1xyXG5cdHRyYW5zaXRpb246IHN0cmluZztcclxuXHQvKipcclxuXHQgKiAgRGV0ZXJtaW5lcyBjc3MtcnVsZSAnd2lkdGgnXHJcbiAgICovXHJcblx0d2lkdGg6IG51bWJlciB8IHN0cmluZztcclxuXHQvKipcclxuXHQgKiAgRGV0ZXJtaW5lcyBjc3MtcnVsZSAncGFkZGluZy1sZWZ0J1xyXG4gICAqL1xyXG5cdHBhZGRpbmdMOiBudW1iZXIgfCBzdHJpbmc7XHJcblx0LyoqXHJcblx0ICogIERldGVybWluZXMgY3NzLXJ1bGUgJ3BhZGRpbmctcmlnaHQnXHJcbiAgICovXHJcblx0cGFkZGluZ1I6IG51bWJlciB8IHN0cmluZztcclxufSJdfQ==
+//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoic3RhZ2UtZGF0YS5tb2RlbC5qcyIsInNvdXJjZVJvb3QiOiJuZzovL25neC1vd2wtY2Fyb3VzZWwtby8iLCJzb3VyY2VzIjpbImxpYi9tb2RlbHMvc3RhZ2UtZGF0YS5tb2RlbC50cyJdLCJuYW1lcyI6W10sIm1hcHBpbmdzIjoiQUFBQTs7R0FFRztBQUNIO0lBQUE7SUFxQkEsQ0FBQztJQUFELGdCQUFDO0FBQUQsQ0FBQyxBQXJCRCxJQXFCQyIsInNvdXJjZXNDb250ZW50IjpbIi8qKlxyXG4gKiBEYXRhIG1vZGVsIGZvciBtYW5hZ2luZyBjbGFzc2VzIG9mIC5vd2wtc3RhZ2UgRE9NIGVsZW1lbnRcclxuICovXHJcbmV4cG9ydCBjbGFzcyBTdGFnZURhdGEge1xyXG5cdC8qKlxyXG5cdCAqIERldGVybWluZXMgY3NzLXJ1bGUgJ3RyYW5zZm9ybSdcclxuXHQgKi9cclxuXHR0cmFuc2Zvcm06IHN0cmluZztcclxuXHQvKipcclxuXHQgKiAgRGV0ZXJtaW5lcyBjc3MtcnVsZSAndHJhbnNpdGlvbidcclxuICAgKi9cclxuXHR0cmFuc2l0aW9uOiBzdHJpbmc7XHJcblx0LyoqXHJcblx0ICogIERldGVybWluZXMgY3NzLXJ1bGUgJ3dpZHRoJ1xyXG4gICAqL1xyXG5cdHdpZHRoOiBudW1iZXIgfCBzdHJpbmc7XHJcblx0LyoqXHJcblx0ICogIERldGVybWluZXMgY3NzLXJ1bGUgJ3BhZGRpbmctbGVmdCdcclxuICAgKi9cclxuXHRwYWRkaW5nTDogbnVtYmVyIHwgc3RyaW5nO1xyXG5cdC8qKlxyXG5cdCAqICBEZXRlcm1pbmVzIGNzcy1ydWxlICdwYWRkaW5nLXJpZ2h0J1xyXG4gICAqL1xyXG5cdHBhZGRpbmdSOiBudW1iZXIgfCBzdHJpbmc7XHJcbn0iXX0=
