@@ -1,24 +1,10 @@
-/**
- * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
- */
 import * as tslib_1 from "tslib";
 import { ErrorHandler, Injectable, isDevMode } from '@angular/core';
 var OwlLogger = /** @class */ (function () {
     function OwlLogger(errorHandler) {
         this.errorHandler = errorHandler;
     }
-    /**
-     * @param {?} value
-     * @param {...?} rest
-     * @return {?}
-     */
-    OwlLogger.prototype.log = /**
-     * @param {?} value
-     * @param {...?} rest
-     * @return {?}
-     */
-    function (value) {
+    OwlLogger.prototype.log = function (value) {
         var rest = [];
         for (var _i = 1; _i < arguments.length; _i++) {
             rest[_i - 1] = arguments[_i];
@@ -27,46 +13,21 @@ var OwlLogger = /** @class */ (function () {
             console.log.apply(console, tslib_1.__spread([value], rest));
         }
     };
-    /**
-     * @param {?} error
-     * @return {?}
-     */
-    OwlLogger.prototype.error = /**
-     * @param {?} error
-     * @return {?}
-     */
-    function (error) {
+    OwlLogger.prototype.error = function (error) {
         this.errorHandler.handleError(error);
     };
-    /**
-     * @param {?} value
-     * @param {...?} rest
-     * @return {?}
-     */
-    OwlLogger.prototype.warn = /**
-     * @param {?} value
-     * @param {...?} rest
-     * @return {?}
-     */
-    function (value) {
+    OwlLogger.prototype.warn = function (value) {
         var rest = [];
         for (var _i = 1; _i < arguments.length; _i++) {
             rest[_i - 1] = arguments[_i];
         }
         console.warn.apply(console, tslib_1.__spread([value], rest));
     };
-    OwlLogger.decorators = [
-        { type: Injectable }
-    ];
-    /** @nocollapse */
-    OwlLogger.ctorParameters = function () { return [
-        { type: ErrorHandler }
-    ]; };
+    OwlLogger = tslib_1.__decorate([
+        Injectable(),
+        tslib_1.__metadata("design:paramtypes", [ErrorHandler])
+    ], OwlLogger);
     return OwlLogger;
 }());
 export { OwlLogger };
-if (false) {
-    /** @type {?} */
-    OwlLogger.prototype.errorHandler;
-}
-//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoibG9nZ2VyLnNlcnZpY2UuanMiLCJzb3VyY2VSb290Ijoibmc6Ly9uZ3gtb3dsLWNhcm91c2VsLW8vIiwic291cmNlcyI6WyJsaWIvc2VydmljZXMvbG9nZ2VyLnNlcnZpY2UudHMiXSwibmFtZXMiOltdLCJtYXBwaW5ncyI6Ijs7Ozs7QUFBQSxPQUFPLEVBQUUsWUFBWSxFQUFFLFVBQVUsRUFBRSxTQUFTLEVBQUUsTUFBTSxlQUFlLENBQUM7QUFFcEU7SUFHRSxtQkFBb0IsWUFBMEI7UUFBMUIsaUJBQVksR0FBWixZQUFZLENBQWM7SUFBRyxDQUFDOzs7Ozs7SUFFbEQsdUJBQUc7Ozs7O0lBQUgsVUFBSSxLQUFVO1FBQUUsY0FBYzthQUFkLFVBQWMsRUFBZCxxQkFBYyxFQUFkLElBQWM7WUFBZCw2QkFBYzs7UUFDNUIsSUFBSSxTQUFTLEVBQUUsRUFBRTtZQUNmLE9BQU8sQ0FBQyxHQUFHLE9BQVgsT0FBTyxvQkFBSyxLQUFLLEdBQUssSUFBSSxHQUFFO1NBQzdCO0lBQ0gsQ0FBQzs7Ozs7SUFFRCx5QkFBSzs7OztJQUFMLFVBQU0sS0FBWTtRQUNoQixJQUFJLENBQUMsWUFBWSxDQUFDLFdBQVcsQ0FBQyxLQUFLLENBQUMsQ0FBQztJQUN2QyxDQUFDOzs7Ozs7SUFFRCx3QkFBSTs7Ozs7SUFBSixVQUFLLEtBQVU7UUFBRSxjQUFjO2FBQWQsVUFBYyxFQUFkLHFCQUFjLEVBQWQsSUFBYztZQUFkLDZCQUFjOztRQUM3QixPQUFPLENBQUMsSUFBSSxPQUFaLE9BQU8sb0JBQU0sS0FBSyxHQUFLLElBQUksR0FBRTtJQUMvQixDQUFDOztnQkFqQkYsVUFBVTs7OztnQkFGRixZQUFZOztJQW9CckIsZ0JBQUM7Q0FBQSxBQWxCRCxJQWtCQztTQWpCWSxTQUFTOzs7SUFFUixpQ0FBa0MiLCJzb3VyY2VzQ29udGVudCI6WyJpbXBvcnQgeyBFcnJvckhhbmRsZXIsIEluamVjdGFibGUsIGlzRGV2TW9kZSB9IGZyb20gJ0Bhbmd1bGFyL2NvcmUnO1xyXG5cclxuQEluamVjdGFibGUoKVxyXG5leHBvcnQgY2xhc3MgT3dsTG9nZ2VyIHtcclxuXHJcbiAgY29uc3RydWN0b3IocHJpdmF0ZSBlcnJvckhhbmRsZXI6IEVycm9ySGFuZGxlcikge31cclxuXHJcbiAgbG9nKHZhbHVlOiBhbnksIC4uLnJlc3Q6IGFueVtdKSB7XHJcbiAgICBpZiAoaXNEZXZNb2RlKCkpIHtcclxuICAgICAgY29uc29sZS5sb2codmFsdWUsIC4uLnJlc3QpO1xyXG4gICAgfVxyXG4gIH1cclxuXHJcbiAgZXJyb3IoZXJyb3I6IEVycm9yKSB7XHJcbiAgICB0aGlzLmVycm9ySGFuZGxlci5oYW5kbGVFcnJvcihlcnJvcik7XHJcbiAgfVxyXG5cclxuICB3YXJuKHZhbHVlOiBhbnksIC4uLnJlc3Q6IGFueVtdKSB7XHJcbiAgICBjb25zb2xlLndhcm4odmFsdWUsIC4uLnJlc3QpO1xyXG4gIH1cclxufVxyXG4iXX0=
+//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoibG9nZ2VyLnNlcnZpY2UuanMiLCJzb3VyY2VSb290Ijoibmc6Ly9uZ3gtb3dsLWNhcm91c2VsLW8vIiwic291cmNlcyI6WyJsaWIvc2VydmljZXMvbG9nZ2VyLnNlcnZpY2UudHMiXSwibmFtZXMiOltdLCJtYXBwaW5ncyI6IjtBQUFBLE9BQU8sRUFBRSxZQUFZLEVBQUUsVUFBVSxFQUFFLFNBQVMsRUFBRSxNQUFNLGVBQWUsQ0FBQztBQUdwRTtJQUVFLG1CQUFvQixZQUEwQjtRQUExQixpQkFBWSxHQUFaLFlBQVksQ0FBYztJQUFHLENBQUM7SUFFbEQsdUJBQUcsR0FBSCxVQUFJLEtBQVU7UUFBRSxjQUFjO2FBQWQsVUFBYyxFQUFkLHFCQUFjLEVBQWQsSUFBYztZQUFkLDZCQUFjOztRQUM1QixJQUFJLFNBQVMsRUFBRSxFQUFFO1lBQ2YsT0FBTyxDQUFDLEdBQUcsT0FBWCxPQUFPLG9CQUFLLEtBQUssR0FBSyxJQUFJLEdBQUU7U0FDN0I7SUFDSCxDQUFDO0lBRUQseUJBQUssR0FBTCxVQUFNLEtBQVk7UUFDaEIsSUFBSSxDQUFDLFlBQVksQ0FBQyxXQUFXLENBQUMsS0FBSyxDQUFDLENBQUM7SUFDdkMsQ0FBQztJQUVELHdCQUFJLEdBQUosVUFBSyxLQUFVO1FBQUUsY0FBYzthQUFkLFVBQWMsRUFBZCxxQkFBYyxFQUFkLElBQWM7WUFBZCw2QkFBYzs7UUFDN0IsT0FBTyxDQUFDLElBQUksT0FBWixPQUFPLG9CQUFNLEtBQUssR0FBSyxJQUFJLEdBQUU7SUFDL0IsQ0FBQztJQWhCVSxTQUFTO1FBRHJCLFVBQVUsRUFBRTtpREFHdUIsWUFBWTtPQUZuQyxTQUFTLENBaUJyQjtJQUFELGdCQUFDO0NBQUEsQUFqQkQsSUFpQkM7U0FqQlksU0FBUyIsInNvdXJjZXNDb250ZW50IjpbImltcG9ydCB7IEVycm9ySGFuZGxlciwgSW5qZWN0YWJsZSwgaXNEZXZNb2RlIH0gZnJvbSAnQGFuZ3VsYXIvY29yZSc7XHJcblxyXG5ASW5qZWN0YWJsZSgpXHJcbmV4cG9ydCBjbGFzcyBPd2xMb2dnZXIge1xyXG5cclxuICBjb25zdHJ1Y3Rvcihwcml2YXRlIGVycm9ySGFuZGxlcjogRXJyb3JIYW5kbGVyKSB7fVxyXG5cclxuICBsb2codmFsdWU6IGFueSwgLi4ucmVzdDogYW55W10pIHtcclxuICAgIGlmIChpc0Rldk1vZGUoKSkge1xyXG4gICAgICBjb25zb2xlLmxvZyh2YWx1ZSwgLi4ucmVzdCk7XHJcbiAgICB9XHJcbiAgfVxyXG5cclxuICBlcnJvcihlcnJvcjogRXJyb3IpIHtcclxuICAgIHRoaXMuZXJyb3JIYW5kbGVyLmhhbmRsZUVycm9yKGVycm9yKTtcclxuICB9XHJcblxyXG4gIHdhcm4odmFsdWU6IGFueSwgLi4ucmVzdDogYW55W10pIHtcclxuICAgIGNvbnNvbGUud2Fybih2YWx1ZSwgLi4ucmVzdCk7XHJcbiAgfVxyXG59XHJcbiJdfQ==
